Extract python_markers.rs from turn.rs (parent #680)
Fifth vertical-slice extraction: move the Python package-marker
materialization helpers out of turn.rs into a new
src/agent/loop_run/python_markers.rs sibling module.

Migrated (all free fns over &mut Agent):

Entry points (pub(super)):
- materialize_python_package_markers_for_external_import — creates
  `__init__.py` markers under the workspace for packages observed in
  verifier stdout/stderr external-import diagnostics.
- materialize_python_package_markers_for_owned_test_imports — same
  materialization but seeded from the owned test artifact list.

Private helper:
- materialize_python_package_marker_candidates — shared per-path
  `create_new` writer + evidence/telemetry record.

The shared helper calls `agent.observe_evidence_from_repo_edit(...)`
(still on impl Agent in turn.rs, pub(super)) to record the marker
creation as evidence. No facade re-export (DR3-001).

Caller updates (verifier_orchestration.rs ×2):
- `agent.materialize_python_package_markers_for_owned_test_imports(...)`
  → `super::python_markers::materialize_python_package_markers_for_owned_test_imports(agent, ...)`
- `agent.materialize_python_package_markers_for_external_import(...)`
  → `super::python_markers::materialize_python_package_markers_for_external_import(agent, ...)`

turn.rs: 5,944 -> 5,883 (-61 LOC)
Cumulative from 39,489: -33,606 LOC (-85.1%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
